Improvements to dataGen script
