+++ server/src/dev-data/seeds/generator.js
@@ -44,3 +44,35 @@ module.exports.generateMockReward = () => ({
     faker.commerce.product()
   ).join(',')
 });
+
+/**
+ * Generate Mock CSV Reward
+ * Creates a fake reward object to be used in seeding the database
+ * @returns
+ */
+module.exports.generateMockCSVReward = (projectID, rewardID) => {
+  // stage our faker data
+  const title = faker.commerce.productName();
+  const pledgeAmount = Math.floor(faker.finance.amount());
+  const description = faker.lorem.paragraph().substring(0, 200);
+  const deliveryMonth = faker.date.month();
+  const deliveryYear = faker.date.future().getFullYear();
+  const shippingType = faker.company.bsAdjective();
+  const rewardQuantity = Math.floor(Math.random() * (500 - 1 + 1)) + 1;
+  const timeLimit = faker.random.number();
+  const projectId = faker.random.number();
+  const rewardItems = Array.from({ length: random.int(1, 6) }, () =>
+    faker.commerce.product()
+  ).join(',')
+  // format each CSV line
+  return `${projectID}, ${rewardID}, ${title},
+    ${pledgeAmount},
+    ${description},
+    ${deliveryMonth},
+    ${deliveryYear},
+    ${shippingType},
+    ${rewardQuantity},
+    ${timeLimit},
+    ${projectId},
+    ${rewardItems}\n`;
+};
